requirements.txt
streamlit
transformers
torch
PyMuPDF
pytesseract
Pillow
textblob
streamlit-lottie
